feat: 서버 예외 처리 추가  - exception 발생 시 내부 에러 응답 코드 리턴  - opcode가 FI 아닌 경우 처리  - 포맷이 잘못된 경우 해당 패킷 삭제  - 클라이언트, 서버 구동 Applicaiton 분리
