+++ build.gradle
@@ -28,6 +28,7 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	implementation group: 'io.netty', name: 'netty-all', version: '4.1.79.Final'
 	implementation 'com.google.code.gson:gson:2.9.1'
+	implementation 'com.google.code.findbugs:jsr305:3.0.2'
 }
 
 tasks.named('test') {
@@ -37,4 +38,8 @@ tasks.named('test') {
 tasks.named('bootBuildImage') {
 	builder = 'paketobuildpacks/builder:tiny'
 	environment = ['BP_NATIVE_IMAGE': 'true']
+}
+
+springBoot {
+	mainClass = "com.nettytransfile.TransFileServerApplication"
 }
